+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/MessageRepository.java
@@ -84,7 +84,7 @@ public class MessageRepository implements MessageContract.Repository {
                                     // 去除没有聊天消息的
                                     continue;
                                 }
-                                tmp.setIm_uid(AppApplication.getmCurrentLoginAuth().getUser_id());
+                                tmp.setIm_uid((int) AppApplication.getmCurrentLoginAuth().getUser_id());
                                 if (tmp.getType() == Conversation.CONVERSATION_TYPE_PRIVATE) { // 单聊
                                     try {
                                         String[] uidsPair = tmp.getUsids().split(",");
@@ -189,7 +189,7 @@ public class MessageRepository implements MessageContract.Repository {
                                          tmp.setLast_message(message);
                                          tmp.setLast_message_time(message.getCreate_time());
                                      }
-                                     tmp.setIm_uid(AppApplication.getmCurrentLoginAuth().getUser_id());
+                                     tmp.setIm_uid((int) AppApplication.getmCurrentLoginAuth().getUser_id());
                                      if (tmp.getType() == Conversation.CONVERSATION_TYPE_PRIVATE) { // 单聊
                                          try {
                                              String[] uidsPair = tmp.getUsids().split(",");
